ether: usb: Fix mismerge in dwc3_remove
Linux moved the pm_runtime_put / disable to the end of the remove,
which we have in our code.  Unfortunately, we also have an extra
pm_runtime_disable left over where it used to be called.

This removes that extra call and gets us inline with how linux
handles pm here.

Change-Id: I3a18beaa713f45e147cfeda94222233442baeee6
Signed-off-by: Joe Maples <joe@frap129.org>