fpc1020: Convert to State Notifier
Thanks to the defer notification time in state notifier, this
fixes the infamous double click effect on screen off. I plan to
extend this to the power button as well to disable use while the fp
sensor is active so that we can use wake and unlock in the rom.

Signed-off-by: Joe Maples <joe@frap129.org>